fix: MatrixOne compatibility and E2E test reliability
- is_duplicate_key_error: match SQLSTATE 23000 or message (sqlx returns SQLSTATE not errno)
- try_claim_lease: CASE WHEN for is_active (MatrixOne returns bool not int)
- list_executions: CAST completed_at AS CHAR for DATETIME→String
- submit_feedback: query agent_events instead of skill_selection_events
- agent_events: add user_feedback_score column
- session_audit: event_type 'turn' → 'user_query' (matches bridge output)
- storage: VARCHAR(36) → VARCHAR(64) for all ID columns; ctx_snapshots extra columns
- team_persistence: shorten builtin team ID prefixes (bt-rev/res/dev)
- memory_health/metrics: derive totals from analyze+hygiene, storage fallback
- chat_stream_handler: bridge-e2e-hooks shortcut for mock LLM in tests
- E2E harness: bootstrap registers mock model; collect_sse_body_text with 5s timeout
- E2E tests: fix event_type in test data, use test_llm_rounds for stream test

+++ rust/crates/runtime/src/turn/agentic_loop_tool_phase.rs
@@ -460,7 +460,7 @@ pub(crate) async fn execute_tool_phase<H: AgenticLoopHost>(
                 let timing = crate::observability_integration::TurnTiming {
                     turn: turn_index as u32,
                     context_assembly_ms: ctx_asm_ms,
-                    ttft_ms: turn_result.ttft_ms.unwrap_or(0) as u64,
+                    ttft_ms: turn_result.ttft_ms.unwrap_or(0),
                     llm_total_ms: total_ms
                         .saturating_sub(ctx_asm_ms)
                         .saturating_sub(tool_exec_ms),
@@ -498,7 +498,7 @@ fn observe_gate_cancelled(
         let timing = crate::observability_integration::TurnTiming {
             turn: turn_index as u32,
             context_assembly_ms: 0,
-            ttft_ms: turn_result.ttft_ms.unwrap_or(0) as u64,
+            ttft_ms: turn_result.ttft_ms.unwrap_or(0),
             llm_total_ms: total_ms,
             tool_execution_ms: 0,
             total_ms,
